Merge feat/backend-core into master via integration branch

+++ backend/src/modules/chapters/chapters.model.ts
@@ -10,7 +10,9 @@ export async function listChapters(
             count(*) OVER ()::int AS total
        FROM corpus.chapters c
        LEFT JOIN corpus.hadiths h ON h.chapter_id = c.chapter_id
-      WHERE c.collection_id = $1
+      -- ::integer so an out-of-smallint-range id matches nothing instead of
+      -- raising PG 22003 (out of range for type smallint).
+      WHERE c.collection_id = $1::integer
       GROUP BY c.chapter_id, c.collection_id, c.seq, c.title_ar
       ORDER BY c.seq
       LIMIT $2 OFFSET $3`,
